[+] Changes in ideology
When I divide project to different source and header files there appear errors when freeing resources.
So for the node and tree classes need to develop copying constructor and destructor.

The binary is works, but when program is ending there is error with double deleting elements->this is error which will be resolved some time that moment.